+++ src/models/common/ohlcv.ts
@@ -1,0 +1,19 @@
+export enum OHLCVPositions {
+    TIME = 0,
+    OPEN = 1,
+    HIGH = 2,
+    LOW = 3,
+    CLOSE = 4,
+    VOLUME = 5,
+}
+
+/**
+ * Used for final kline events
+ *
+ * Timestamp
+ * Open
+ * High
+ * Low
+ * Close
+ */
+export type OHLCV = [number, number, number, number, number, number];